chore: rename device_driver to driver to avoid dup
