+++ crates/device/src/lib.rs
@@ -1,7 +1,7 @@
 #![no_std]
 
 pub mod common;
-pub mod device_driver;
+pub mod driver;
 pub mod gpio;
 pub mod mailbox;
 pub mod mini_uart;
